add dialogs for term property content

+++ app/Http/Controllers/TermController.php
@@ -2,6 +2,7 @@
 namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Term;
+use App\TermProperty;
 use App\Glossary;
 use App\Ontology;
 use App\Status;
@@ -104,6 +105,7 @@ class TermController extends Controller
 
 		//validate input form
 		$this->validate($request, [
+			'term_id' => 'required',
 			'term_name' => 'required|min:3',
 			'term_description' => 'required',
 			'glossary_id' => 'required',
@@ -121,7 +123,22 @@ class TermController extends Controller
 				$ontology->relation_id = $relation['relation_id'];
 				$ontology->object_id = $relation['object_id'];
 				$ontology->status_id = $request->input('status_id');
+				$ontology->created_by = Auth::user()->id;
 				$ontology->save();
+			}
+		}
+
+		//detele existing term properties
+		TermProperty::where('term_id', $request->input('term_id'))->delete();
+
+		foreach($request->input('Properties') as $property) {
+			if (!empty($property['property_name'])) {
+				$termproperty = new TermProperty;
+				$termproperty->term_id = $request->input('term_id');
+				$termproperty->property_name = $property['property_name'];
+				$termproperty->property_value = $property['property_value'];
+				$termproperty->created_by = Auth::user()->id;
+				$termproperty->save();
 			}
 		}
 
@@ -136,16 +153,33 @@ class TermController extends Controller
 			abort(403, 'Unauthorized action.');
 		}
 
+		//check if id property exists
+		if (!$term->id) {
+			abort(403, 'This term no longer exists in the database.');
+		}
+
+		//delete ontology
+		Ontology::where('subject_id', $term->id)->delete();
+		Ontology::where('object_id', $term->id)->delete();
+
 		$term->delete();
-		return Redirect::to('/terms?letter=' . substr($term->term_name, 0, 1))->with('message', 'Term deleted.');
+		return Redirect::to('/terms?letter=' . substr($term->term_name, 0, 1))->with('message', 'Term and related ontology deleted.');
 	}
 
 	public function apiIndex(Request $request)
 	{
 		if ($request->has('search')) {
 			$terms = Term::where('term_name', 'like', '%' . $request->input('search') . '%')->orderBy('term_name', 'asc')->orWhere(function ($query) use ($request) {
-						$query->where('term_description', 'like', '%' . $request->input('search') . '%')->orderBy('term_name', 'asc');
-					})->get();
+				$query->where('term_description', 'like', '%' . $request->input('search') . '%')->orderBy('term_name', 'asc');
+			})->get();
+		} elseif ($request->has('property_name') && $request->has('property_value')) {
+			$terms = Term::whereHas('properties', function ($query) use ($request) {
+				$query->where('property_name', 'like', '%' . $request->input('property_name') . '%')->where('property_value', 'like', '%' . $request->input('property_value') . '%');
+			})->orderBy('term_name', 'asc')->get();
+		} elseif ($request->has('property_name')) {
+			$terms = Term::whereHas('properties', function ($query) use ($request) {
+				$query->where('property_name', 'like', '%' . $request->input('property_name') . '%');
+			})->orderBy('term_name', 'asc')->get();
 		} else {
 			$terms = Term::orderBy('term_name', 'asc')->get();
 		}
@@ -153,16 +187,18 @@ class TermController extends Controller
 		//return response()->json($terms);
 
 		$result = array();
-		foreach ($terms as $term) {
-			array_push($result, array(
-					"id"  => round($term->id),
-					"term_name"  => $term->term_name,
-					"term_description" => preg_replace("/^(.{250})([^\.]*\.)(.*)$/", "\\1\\2", $term->term_description),
-					"glossary_name" => $term->glossary->glossary_name,
-					"value" => $term->term_name,
-					"tokens" => array($term->term_name)
-				)
-			);
+		if (!empty($terms)) {
+			foreach ($terms as $term) {
+				array_push($result, array(
+						"id"  => round($term->id),
+						"term_name"  => $term->term_name,
+						"term_description" => preg_replace("/^(.{250})([^\.]*\.)(.*)$/", "\\1\\2", $term->term_description),
+						"glossary_name" => $term->glossary->glossary_name,
+						"value" => $term->term_name,
+						"tokens" => array($term->term_name)
+					)
+				);
+			}
 		}
 
 		return response()->json($result);
